% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{confirmar_alteracao}
\alias{confirmar_alteracao}
\title{Confirma a alteração no módulo dependente}
\usage{
confirmar_alteracao(historico_alteracoes, pagina_id)
}
\arguments{
\item{historico_alteracoes}{}

\item{pagina_id}{}
}
\value{
list
}
\description{
Confirma a alteração no módulo dependente
}
