% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create_bd}
\alias{create_bd}
\title{Cria o banco de dados}
\usage{
create_bd(con)
}
\arguments{
\item{con}{}
}
\value{
data base created
}
\description{
Cria o banco de dados
}
\examples{
create_bd(con)
}
