% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fishApp_bd.R
\name{create_fishApp_bd}
\alias{create_fishApp_bd}
\title{Cria o banco de dados}
\usage{
create_fishApp_bd(con)
}
\arguments{
\item{con}{}
}
\value{
data base created
}
\description{
Cria o banco de dados
}
\examples{
create_fishApp_bd(con)
}
