% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fishApp_tables.R
\name{create_fishApp_tables}
\alias{create_fishApp_tables}
\title{Cria as tabelas do banco de dados}
\usage{
create_fishApp_tables(con)
}
\arguments{
\item{con}{}
}
\value{
tables
}
\description{
Cria as tabelas do banco de dados
}
\examples{
create_fishApp_tables(con)
}
