% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{get_bases_fixas}
\alias{get_bases_fixas}
\title{Buscar bases_fixas}
\usage{
get_bases_fixas(con, projeto_id)
}
\arguments{
\item{con}{}
}
\value{
data_frame
}
\description{
Buscar bases_fixas
}
\examples{
get_bases_fixas(con)
}
