% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{get_equipamentos}
\alias{get_equipamentos}
\title{Buscar equipamentos}
\usage{
get_equipamentos(con, projeto_id)
}
\arguments{
\item{con}{}
}
\value{
data_frame
}
\description{
Buscar equipamentos
}
\examples{
get_equipamentos(con)
}
