% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{get_pessoas_projetos}
\alias{get_pessoas_projetos}
\title{Buscar pessoas}
\usage{
get_pessoas_projetos(con)
}
\arguments{
\item{con}{}
}
\value{
data_frame
}
\description{
Buscar pessoas
}
\examples{
get_projetos(con)
}
