% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{get_tipo_equipamento}
\alias{get_tipo_equipamento}
\title{Buscar get_tipo_equipamento}
\usage{
get_tipo_equipamento(con)
}
\arguments{
\item{con}{}
}
\value{
data_frame
}
\description{
Buscar get_tipo_equipamento
}
\examples{
get_tipo_equipamento(con)
}
