% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{get_tipo_local}
\alias{get_tipo_local}
\title{Buscar tipo_local}
\usage{
get_tipo_local(con)
}
\arguments{
\item{con}{}
}
\value{
data_frame
}
\description{
Buscar tipo_local
}
\examples{
get_tipo_local(con)
}
