% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{informar_alteracao}
\alias{informar_alteracao}
\title{Comunica alterações para modulos dependentes}
\usage{
informar_alteracao(historico_alteracoes, pagina_id)
}
\arguments{
\item{historico_alteracoes}{}

\item{pagina_id}{}
}
\value{
list
}
\description{
Comunica alterações para modulos dependentes
}
